Split goals into high and low priority-ish
The actual values have yet to be assigned, but all the information
that's at hand makes it fairly easy to do so. The supernodes have a
boolean variable that keeps track of it being a goal supernode or not,
as well as a variable that keeps track of if the node is absorbed or not
- which is what dictates if it's high priority or not. absorbed = high
priority. not absorbed = low priority.